Simplified image classification example, updated readme and in-code docs.
